ingest-daemon/pyproject.toml
[tool.poetry]
name = "ingest-daemon"
version = "0.1.0"
description = ""
authors = ["ov3rwrite <ilyabelykh123@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
aio-pika = "^9.4.0"
boto3 = "^1.34.33"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
